Refactor ImagePullPolicy usage to use the correct namespace and add phpstan configuration
