+++ src/Containers/Container.php
@@ -5,6 +5,7 @@ namespace Testcontainers\Containers;
 use Testcontainers\Containers\PortStrategy\PortStrategy;
 use Testcontainers\Containers\StartupCheckStrategy\StartupCheckStrategy;
 use Testcontainers\Containers\Types\HostToIp;
+use Testcontainers\Containers\Types\ImagePullPolicy;
 use Testcontainers\Containers\WaitStrategy\WaitStrategy;
 
 /**
